Add header and refactor checksigs into challenges
