+++ pkg/ca/fulcioca/fulcioca_test.go
@@ -24,7 +24,7 @@ import (
 	"crypto/sha256"
 	"testing"
 
-	"github.com/sigstore/fulcio/pkg/ca/cautils"
+	"github.com/sigstore/fulcio/pkg/challenges"
 )
 
 func failErr(t *testing.T, err error) {
@@ -39,7 +39,7 @@ func TestCheckSignatureECDSA(t *testing.T) {
 	failErr(t, err)
 
 	email := "test@gmail.com"
-	if err := cautils.CheckSignature(&priv.PublicKey, []byte("foo"), email); err == nil {
+	if err := challenges.CheckSignature(&priv.PublicKey, []byte("foo"), email); err == nil {
 		t.Fatal("check should have failed")
 	}
 
@@ -47,12 +47,12 @@ func TestCheckSignatureECDSA(t *testing.T) {
 	signature, err := priv.Sign(rand.Reader, h[:], crypto.SHA256)
 	failErr(t, err)
 
-	if err := cautils.CheckSignature(&priv.PublicKey, signature, email); err != nil {
+	if err := challenges.CheckSignature(&priv.PublicKey, signature, email); err != nil {
 		t.Fatal(err)
 	}
 
 	// Try a bad email but "good" signature
-	if err := cautils.CheckSignature(&priv.PublicKey, signature, "bad@email.com"); err == nil {
+	if err := challenges.CheckSignature(&priv.PublicKey, signature, "bad@email.com"); err == nil {
 		t.Fatal("check should have failed")
 	}
 }
@@ -62,7 +62,7 @@ func TestCheckSignatureRSA(t *testing.T) {
 	failErr(t, err)
 
 	email := "test@gmail.com"
-	if err := cautils.CheckSignature(&priv.PublicKey, []byte("foo"), email); err == nil {
+	if err := challenges.CheckSignature(&priv.PublicKey, []byte("foo"), email); err == nil {
 		t.Fatal("check should have failed")
 	}
 
@@ -70,12 +70,12 @@ func TestCheckSignatureRSA(t *testing.T) {
 	signature, err := priv.Sign(rand.Reader, h[:], crypto.SHA256)
 	failErr(t, err)
 
-	if err := cautils.CheckSignature(&priv.PublicKey, signature, email); err != nil {
+	if err := challenges.CheckSignature(&priv.PublicKey, signature, email); err != nil {
 		t.Fatal(err)
 	}
 
 	// Try a bad email but "good" signature
-	if err := cautils.CheckSignature(&priv.PublicKey, signature, "bad@email.com"); err == nil {
+	if err := challenges.CheckSignature(&priv.PublicKey, signature, "bad@email.com"); err == nil {
 		t.Fatal("check should have failed")
 	}
 }
